Updated the processing code to handle the added entry time statistic for GeoEntries.
I also found a bug. Coordinates were stored as two ints. This was changed to be stored as two floats
so that the results will have more precision.

+++ trunk/utils/LocationDeltaStatistics/lib/Coordinate.rb
@@ -2,7 +2,7 @@ class Coordinate
     attr_accessor :x, :y
 
     def initialize(x = 0,  y = 0)
-        @x, @y = x.to_i, y.to_i
+        @x, @y = x.to_f, y.to_f
     end
 
     def distance(coordinate)
